Allow configuring public host in demos
Prior to this PR, `npm run demo` created an app where hot-reloading
did not work if the app was behind a proxy, because the browser client
looked for the backend at localhost, rather than using
window.location. To work at all, we disabled the host check, which, in
principle, opened the backend to phishing attacks [1].

Now, we allow configuring the host header for the frontend using the
DEMO_PUBLIC_HOST environment variable. If that variable is set, it
will be used by the frontend and by the host check in the backend. If
not, we fall back to the webpack default (localhost:<port>, in most
occasions).

[1] These would not work on SherlockML because we have an additional
layer of security that verifies the host header.