Run npm install when publishing
As of PR 105 (allow publishing to npm), we no longer ran
`npm install` when publishing because I was under the (mistaken)
assumption that `npm publish` also ran `npm install`.

This was probably not apparent during testing because we were re-using
node_modules from the last successful installation.

The task supposed to clean the node modules was not removing the right
directory. Since being able to cache node_modules is mostly a good
thing (for speed etc.), I removed it.