Fix DropdownMenu dismissal
Previously DropdownMenuItems that were being used as internal links did
not correctly dismiss the dropdown menu. This is because the internal
links call preventDefault on the click event to stop the page from
refreshing. However this would stop the event from reaching the
underlying reactstrap component which would handle the dismissal.

This fix checks if the DropdownMenuItem is being used as an internal
link, and if so takes care of dismissing the DropdownMenu itself.

To give users even more control, DropdownMenuItem now has a new prop
toggle, True by default, which determines whether clicking on the item
dismisses the menu.