Call setProps only if component is not mounted
Previously we were getting errors due to uncancelled timers not getting
cleaned up properly when the component unmounted. This resulted in a
delayed call to setProps and an error.

Unfortunately cancelling the timeouts directly appears to be
challenging, so this change introduces a simple `isMountedRef` which is
used a guard on calls to setProps.